refactor: extract command utilities
